Tab system implemented with 2 menus + Close tab
The unit and clicker upgrade menus can now be navigated to by selecting their corresponding tab. The 1st tab for the quest/clicker upgrade menu and the second tab for the unit menu. Third and fourth tabs have been page for the upcoming node and voyage menu. There is a fifth tab that is tinted red that does not have any menu attached to it and closes other menus when pressed.